+++ classes/Core.php
@@ -104,9 +104,13 @@ class Core {
     //TO DO calculate dimensions, weight
     public function get_parcels($order = false) {
         global $woocommerce;
+        $config = $this->get_config();
+        $send_as_one = $config['send_as_one'] ?? 'no';
+
         $product = new Product();
-        $product->set_config($this->get_config());
+        $product->set_config($config);
         $parcels = [];
+        $parcel_objects = [];
         if ($order) {
             $items = $order->get_items();
         } else {
@@ -139,7 +143,31 @@ class Core {
             $parcel->setHeight($product_height);
             $parcel->setWidth($product_width);
             $parcel->setLength($product_length);
+            $parcel_objects[] = $parcel;
             $parcels[] = $parcel->generateParcel();
+        }
+
+        if ($send_as_one == 'yes') {
+            $main_parcel = new Parcel();
+            $total_weight = 0;
+            $total_volume = 0;
+            $total_amount = 1;
+            foreach ($parcel_objects as $parcel) {
+                try {
+                    $parcel_data = $parcel->__toArray();
+                    $total_weight += $parcel_data['weight'] * $parcel_data['amount'];
+                    $total_volume += $parcel_data['x'] * $parcel_data['y'] * $parcel_data['z'] * $parcel_data['amount'];
+                } catch (\Throwable $e) {
+
+                }
+            }
+            $cube_size = ceil($total_volume**(1/3));
+            $main_parcel->setAmount(1);
+            $main_parcel->setUnitWeight($total_weight);
+            $main_parcel->setHeight($cube_size);
+            $main_parcel->setWidth($cube_size);
+            $main_parcel->setLength($cube_size);
+            return [$main_parcel->generateParcel()];
         }
         return $parcels;
     }
@@ -222,26 +250,39 @@ class Core {
 
     public function sort_offers(&$offers) {
         $edited_offers = array();
-        $grouped = $this->group_offers($offers);
 
-        foreach ($grouped as $group => $grouped_offers) {
-            $sort_by = $this->config[$group . '_sort_by'] ?? "default";
-            if ($sort_by == "fastest") {
-                usort($grouped[$group], function ($v, $k) {
+        $main_sort_by = $this->config['sort_by'] ?? "default";
+        if ($main_sort_by == 'default') {
+            $grouped = $this->group_offers($offers);
+
+            foreach ($grouped as $group => $grouped_offers) {
+                $sort_by = $this->config[$group . '_sort_by'] ?? "default";
+                if ($sort_by == "fastest") {
+                    usort($grouped[$group], function ($v, $k) {
+                        return $this->get_offer_delivery($k) <= $this->get_offer_delivery($v);
+                    });
+                } elseif ($sort_by == "cheapest") {
+                    usort($grouped[$group], function ($v, $k) {
+                        return $k->price <= $v->price;
+                    });
+                }
+
+                foreach ($grouped[$group] as $offer) {
+                    $edited_offers[] = $offer;
+                }
+            }
+            $offers = $edited_offers;
+        } else {
+            if ($main_sort_by == "fastest") {
+                usort($offers, function ($v, $k) {
                     return $this->get_offer_delivery($k) <= $this->get_offer_delivery($v);
                 });
-            } elseif ($sort_by == "cheapest") {
-                usort($grouped[$group], function ($v, $k) {
+            } elseif ($main_sort_by == "cheapest") {
+                usort($offers, function ($v, $k) {
                     return $k->price <= $v->price;
                 });
             }
-
-            foreach ($grouped[$group] as $offer) {
-                $edited_offers[] = $offer;
-            }
         }
-
-        $offers = $edited_offers;
     }
 
     private function group_offers($offers)